requirements.txt
django=5.2.6
django-jalali
django-jalali-date
pillow
python-decouple
ghasedak-sms
mysqlclient